# Pool block and pool index size performance improvements