--- a/lib/gbenchmark/.clang-format
+++ b/lib/gbenchmark/.clang-format
@@ -1,0 +1,5 @@
+---
+Language:        Cpp
+BasedOnStyle:  Google
+...
+