--- a/lib/gbenchmark/gbenchmark-all.cc
+++ b/lib/gbenchmark/gbenchmark-all.cc
@@ -1,0 +1,16 @@
+#include "benchmark/benchmark.h"
+
+// The following lines pull in the real gbenchmark *.cc files.
+#include "src/benchmark.cc"
+#include "src/benchmark_register.cc"
+#include "src/colorprint.cc"
+#include "src/commandlineflags.cc"
+#include "src/complexity.cc"
+#include "src/console_reporter.cc"
+#include "src/csv_reporter.cc"
+#include "src/json_reporter.cc"
+#include "src/reporter.cc"
+#include "src/sleep.cc"
+#include "src/string_util.cc"
+#include "src/sysinfo.cc"
+#include "src/timers.cc"