--- a/lib/gbenchmark/tools/gbench/__init__.py
+++ b/lib/gbenchmark/tools/gbench/__init__.py
@@ -1,0 +1,8 @@
+"""Google Benchmark tooling"""
+
+__author__ = 'Eric Fiselier'
+__email__ = 'eric@efcs.ca'
+__versioninfo__ = (0, 5, 0)
+__version__ = '.'.join(str(v) for v in __versioninfo__) + 'dev'
+
+__all__ = []