* Modify `pool_block_insert` to insert a free block of equal size
  at beginning instead of end of list of other blocks of equal size.
* Modify `ponyint_pool_index` to remove a branch. @Praetonus can you review this? I think it's better (or no worse) at least on x86_64 but I'm not sure about ARM. I can undo this change if needed.

NOTE: This PR builds on top of PR #1565 that added benchmarks and includes that commit (which should become a no-op if this is applied after that one is merged). The changes for this PR (excluding the commit adding the benchmarks) is only in `src/libponyrt/mem/pool.c`.

Old performance:

```
Benchmark                                              Time           CPU Iterations
------------------------------------------------------------------------------------
PoolBench/pool_index/1                                 2 ns          2 ns  349320441    466.69M items/s
PoolBench/pool_index/3                                 2 ns          2 ns  339153909   456.103M items/s
PoolBench/pool_index/9                                 2 ns          2 ns  331079844   466.057M items/s
PoolBench/pool_index/27                                2 ns          2 ns  348953174   460.689M items/s
PoolBench/pool_index/81                                2 ns          2 ns  303335752   386.927M items/s
PoolBench/pool_index/243                               2 ns          2 ns  276376835   390.288M items/s
PoolBench/pool_index/729                               2 ns          2 ns  280300298   389.146M items/s
PoolBench/pool_index/2.13574k                          2 ns          2 ns  292997048   382.799M items/s
PoolBench/pool_index/6.40723k                          2 ns          2 ns  289336599   404.284M items/s
PoolBench/pool_index/19.2217k                          2 ns          2 ns  299981950   399.787M items/s
PoolBench/pool_index/57.665k                           2 ns          2 ns  292836323    395.01M items/s
PoolBench/pool_index/172.995k                          2 ns          2 ns  294481784   410.526M items/s
PoolBench/pool_index/518.985k                          2 ns          2 ns  279430450    401.86M items/s
PoolBench/pool_index/1024k                             3 ns          3 ns  254565937   367.271M items/s
PoolBench/POOL_ALLOC                                 246 ns        248 ns    2848839   3.84826M items/s
PoolBench/POOL_ALLOC_multiple                      26045 ns      26222 ns      27701   3.63689M items/s
PoolBench/POOL_FREE                                  243 ns        245 ns    2773275   3.89873M items/s
PoolBench/POOL_FREE_multiple                       25462 ns      25615 ns      27450    3.7231M items/s
PoolBench/pool_alloc_size                            249 ns        251 ns    2750278   3.79614M items/s
PoolBench/pool_alloc_size_multiple                170088 ns     170247 ns       4103   573.615k items/s
PoolBench/pool_free_size                            3602 ns       3610 ns     186307   270.527k items/s
PoolBench/pool_free_size_multiple                 174111 ns     174932 ns       3906   558.252k items/s
```

New performance:

```
Benchmark                                              Time           CPU Iterations
------------------------------------------------------------------------------------
PoolBench/pool_index/1                                 2 ns          2 ns  380167950    520.17M items/s
PoolBench/pool_index/3                                 2 ns          2 ns  382537527   523.337M items/s
PoolBench/pool_index/9                                 2 ns          2 ns  382398791   526.296M items/s
PoolBench/pool_index/27                                2 ns          2 ns  381535442   527.559M items/s
PoolBench/pool_index/81                                2 ns          2 ns  325839606    448.56M items/s
PoolBench/pool_index/243                               2 ns          2 ns  327491536   449.827M items/s
PoolBench/pool_index/729                               2 ns          2 ns  320849552   448.942M items/s
PoolBench/pool_index/2.13574k                          2 ns          2 ns  326125609   447.817M items/s
PoolBench/pool_index/6.40723k                          2 ns          2 ns  321929585   443.727M items/s
PoolBench/pool_index/19.2217k                          2 ns          2 ns  324905773   449.437M items/s
PoolBench/pool_index/57.665k                           2 ns          2 ns  325316638   451.658M items/s
PoolBench/pool_index/172.995k                          2 ns          2 ns  327863003   443.843M items/s
PoolBench/pool_index/518.985k                          2 ns          2 ns  321915232   446.399M items/s
PoolBench/pool_index/1024k                             2 ns          2 ns  328150186   451.759M items/s
PoolBench/POOL_ALLOC                                 222 ns        224 ns    3228044    4.2543M items/s
PoolBench/POOL_ALLOC_multiple                      22332 ns      22516 ns      30873   4.23557M items/s
PoolBench/POOL_FREE                                  218 ns        220 ns    3198945   4.34186M items/s
PoolBench/POOL_FREE_multiple                       21868 ns      22083 ns      31458    4.3186M items/s
PoolBench/pool_alloc_size                            218 ns        220 ns    3186569   4.34472M items/s
PoolBench/pool_alloc_size_multiple                 26458 ns      26741 ns      26186   3.56637M items/s
PoolBench/pool_free_size                             221 ns        223 ns    3147969   4.28029M items/s
PoolBench/pool_free_size_multiple                  24443 ns      24728 ns      28923   3.85667M items/s
```