@jemc 2017-02-08
@pdtwonotes any chance you might want to run these benchmarks (before and after) on your ARM setup? :rabbit2: