@pdtwonotes 2017-02-09
I pulled dipinhora:PR_pool_perf and tried to build it on ARM.  I got these errors, which are in hash.cc.  It looks like a 32 vs 64 problem.
```
benchmark/libponyrt/ds/hash.cc: In member function ‘virtual void HashMapBench_HashResize_Benchmark::BenchmarkCase(benchmark::State&)’:
benchmark/libponyrt/ds/hash.cc:139:125: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 2 has type ‘size_t {aka unsigned int}’ [-Werror=format=]
 ize: %lu, old_size: %lu\n", testmap_size(&_map), _map.contents.size, old_size);
                                                                              ^
benchmark/libponyrt/ds/hash.cc:139:125: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 3 has type ‘size_t {aka unsigned int}’ [-Werror=format=]
benchmark/libponyrt/ds/hash.cc:139:125: error: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 4 has type ‘size_t {aka unsigned int}’ [-Werror=format=]
cc1plus: all warnings being treated as errors
Makefile:620: recipe for target 'build/release/obj/benchmarks/libponyrt/ds/hash.o' failed
make: *** [build/release/obj/benchmarks/libponyrt/ds/hash.o] Error 1
make: *** Waiting for unfinished jobs....
111