@dipinhora 2017-02-09
@pdtwonotes thanks for trying to run the benchmarks. It seems my printf debug statements before SkipWithError in the benchmarks are the culprit.

I'll let you know once I've gotten them fixed and verified.