@pdtwonotes 2017-02-09
Changing %lu to %zu fixed that problem.  But I need instructions on how to actually execute the benchmark.