@dipinhora 2017-02-09
@pdtwonotes `make benchmark` will automatically build and run all the benchmarks (including the ones for the PoolBench, HashMapBench, and HeapBench).

If you want to instead run only the PoolBench benchmarks you can run the following after stopping the full benchmarks run that `make benchmark` would try and run:

```
./build/release/libponyrt.benchmarks --benchmark_filter=PoolBench
```

Also, according to this SO at http://stackoverflow.com/a/2524673, `%zu` is not entirely portable and not officially defined in the C++ standard until C++1x. I'll switch things over to using `cout` later today or tomorrow when I have time to spend on this again since the debug output only occurs after we've hit an error condition and performance doesn't matter much.