@pdtwonotes 2017-02-09
Here is what I get on RPi3.  The CPU speed is clearly off.   Now, how do I get the "before" numbers?
```
release> ./libponyrt.benchmarks --benchmark_filter=PoolBench
Run on (4 X 38.4 MHz CPU s)
2017-02-09 12:48:27
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
Benchmark                                   Time           CPU Iterations
-------------------------------------------------------------------------
PoolBench/pool_index/1                     13 ns         13 ns   48442795   76.2355M items/s
PoolBench/pool_index/3                     13 ns         13 ns   55955469   76.2348M items/s
PoolBench/pool_index/9                     13 ns         13 ns   55940073    76.227M items/s
PoolBench/pool_index/27                    13 ns         13 ns   55931670   76.2355M items/s
PoolBench/pool_index/81                    17 ns         17 ns   41965653   57.1768M items/s
PoolBench/pool_index/243                   17 ns         17 ns   41967370   57.1125M items/s
PoolBench/pool_index/729                   17 ns         17 ns   41968812   57.1747M items/s
PoolBench/pool_index/2.13574k              17 ns         17 ns   41966689   57.1744M items/s
PoolBench/pool_index/6.40723k              17 ns         17 ns   41968602   57.1774M items/s
PoolBench/pool_index/19.2217k              17 ns         17 ns   41964894    57.177M items/s
PoolBench/pool_index/57.665k               17 ns         17 ns   41968641   57.1766M items/s
PoolBench/pool_index/172.995k              17 ns         17 ns   41966060   57.1761M items/s
PoolBench/pool_index/518.985k              17 ns         17 ns   41967973   57.1771M items/s
PoolBench/pool_index/1024k                 17 ns         17 ns   41967475   57.1755M items/s
PoolBench/POOL_ALLOC                     1487 ns       1499 ns     466153   651.303k items/s
PoolBench/POOL_ALLOC_multiple          155415 ns     156603 ns       4474   623.593k items/s
PoolBench/POOL_FREE                      1514 ns       1517 ns     461974    643.92k items/s
PoolBench/POOL_FREE_multiple           152130 ns     152868 ns       4582   638.828k items/s
PoolBench/pool_alloc_size                1501 ns       1513 ns     462307   645.358k items/s
PoolBench/pool_alloc_size_multiple     164651 ns     167447 ns       4178   583.208k items/s
PoolBench/pool_free_size                 1522 ns       1532 ns     457019   637.545k items/s
PoolBench/pool_free_size_multiple      157294 ns     158419 ns       4416   616.444k items/s
```