@dipinhora 2017-02-09
@pdtwonotes in order to get the before number you'd need to revert the last commit using `git revert 550b7ba4ca6e299b5c7c823ee76aa4ce62d52ea4` and then recompile and re-run the benchmarks.