@dipinhora 2017-02-09
Also, I just noticed your comment about the cpu speed and scaling. You can do the following to control those:

```
# cpu frequency related stuff
cpupower -c all frequency-set -g performance --min $(cat /sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_max_freq) --max $(cat /sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_max_freq) # ensure cpu runs at max speed (note this is bad for lowering power usage)

# disable turbo mode (cpu scaling)
# add the following to /boot/config.txt and reboot
force_turbo=0 # disable turbo mode for cpu
```